get rid of ciclic variable dependencies
