+++ core/include/scipp/core/tag_util.h
@@ -4,7 +4,6 @@
 /// @author Simon Heybrock
 #ifndef TAG_UTIL_H
 #define TAG_UTIL_H
-#include "scipp/core/dataset.h"
 
 namespace scipp::core {
 
@@ -46,16 +45,6 @@ template <class... Ts> struct CallDType {
 template <class... Ts>
 struct CallDTypeWithSparse : public CallDType<Ts..., sparse_container<Ts>...> {
 };
-
-template <template <class> class Callable, class... Args>
-auto apply(const DType dtype, Args &&... args) {
-  return callDType<Callable>(
-      std::tuple<double, float, int64_t, int32_t, std::string, bool,
-                 sparse_container<double>, sparse_container<float>,
-                 sparse_container<int64_t>, DataArray, Dataset,
-                 Eigen::Vector3d>{},
-      dtype, std::forward<Args>(args)...);
-}
 
 } // namespace scipp::core
