+++ core/include/scipp/core/variable.h
@@ -18,6 +18,7 @@
 #include "scipp/core/except.h"
 #include "scipp/core/slice.h"
 #include "scipp/core/string.h"
+#include "scipp/core/tag_util.h"
 #include "scipp/core/values_and_variances_strong_types.h"
 #include "scipp/core/variable_view.h"
 #include "scipp/units/unit.h"
