+++ core/string.cpp
@@ -150,6 +150,16 @@ template <class T> struct VariancesToString {
   }
 };
 
+template <template <class> class Callable, class... Args>
+auto apply(const DType dtype, Args &&... args) {
+  return callDType<Callable>(
+      std::tuple<double, float, int64_t, int32_t, std::string, bool,
+                 sparse_container<double>, sparse_container<float>,
+                 sparse_container<int64_t>, DataArray, Dataset,
+                 Eigen::Vector3d>{},
+      dtype, std::forward<Args>(args)...);
+}
+
 template <class Key, class Var>
 auto format_variable(const Key &key, const Var &variable,
                      const Dimensions &datasetDims = Dimensions()) {
